pyproject.toml
[tool.poetry]
name = "covid19-bot"
version = "0.1.0"
description = "Special Discord bot made for the Coronavirus covid19"
authors = ["takitsu21"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-decouple = "^3.4"
matplotlib = "^3.4.2"
aiomysql = "^0.0.21"
sentry-sdk = "^1.1.0"
aiohttp = "^3.7.4"
psutil = "^5.8.0"
"discord.py" = "^1.7.3"
discord-py-slash-command = "^2.1.0"
topggpy = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
